Add truncating cmap format4
Windows does not recognize the font
if both of the following two conditions are met.

Fill in gaps in the cmap table format4.
The font has a GPOS table.

Filling in the gaps can keep the size of format4 within the limits,
but it is incompatible with GPOS tables.

So, instead of reducing the size by filling in the gaps,
this commit makes the generater can truncate the table
by removing CJK Unified Ideographs (U+4E00..U+9FFF)
and CJK Compatibility Ideographs (U+F900..U+FAFF)
according to the following document.

https://github.com/adobe-type-tools/Adobe-KR

Currently, {Gothic|Mincho}KR and GothicTW exceed the format4 size,
so this script truncates the table.