Fix `vrt2` to not be a subset of `vert`
https://github.com/trueroad/HaranoAjiFonts/issues/6

The entry for Tag: 'vrt2' feature
in the OpenType specification version 1.8.4 says
"Note that these are a superset of the glyphs covered in the 'vert' table."
https://docs.microsoft.com/en-us/typography/opentype/spec/features_uz#tag-vrt2

However, some glyphs (e.g. U+FF1A 'FULLWIDTH COLON') were covered
by the 'vert' feature, but not by the 'vrt2' feature.
This meant that 'vrt2' is not a superset of 'vert', but a subset.

This is because `vrt2` in Source Han Sans and Source Han Serif
is a subset of `vert`.
https://github.com/adobe-fonts/source-han-sans/issues/282

Source Han Sans and Source Han Serif have
multiple lookup tables for `vert`.
However, the entry for Tag: 'vrt2' feature
in the OpenType specification version 1.8.4 says
"the 'vrt2' feature must be present in the GSUB table,
it must comprises a single lookup of LookupType 1 and LookupFlag 0,
and the lookup must have a single subtable."
https://docs.microsoft.com/en-us/typography/opentype/spec/features_uz#tag-vrt2

Therefore, `vrt2` of Source Han Sans and Source Han Serif contain
only one of several `vert` lookup tables,
which makes `vrt2` a subset of `vert`.

This commit fixes it by integrating lookup tables of `vert`.