Add copying CID+2510 to CID+13463 for HaranoAjiGothic fonts
In SourceHanSans 2.003, uni4E08uE0101-JP glyph,
which was previously mapped to CID+13463 in HaranoAjiGothic 20210130,
is removed.
In old SourceHanSans 2.002, the shape of the uni4E08uE0101-JP glyph
was almost identical to uni4E08-CN glyph mapped to the CID+2510.

This commit makes that the generator copies CID+2510 to CID+13463
when HaranoAjiGothic is generated from SourceHanSansJP.