Add `--jisx0208` option to check_table_CMap
2004-H and 2004-V contain characters that are not in JIS X 0208
(i.e. in only JIS X 0213).
However, the Japanese characters officially supported
by pTeX are only in JIS X 0208.
In other words, characters in JIS X 0213 that are not in JIS X 0208
cannot officially be used in pTeX.

So, this commit adds `--jisx0208` option to check_table_CMap
so that it can only check JIS X 0208 coverage in 2004-H and 2004-V.