Fix an issue where prefer Unicode that does not exist in CMap
When multiple Unicode code points are mapped to the same CID
in the cmap table of SourceHan,
the code point that did not exist in CMap file might be preferred.
This caused some glyphs to be missing in Harano Aji Fonts.

See https://github.com/trueroad/HaranoAjiFontsK1/issues/1

This commit fixes the missing glyphs by Unicode code points
that do not exist in the CMap file are not preferred.