Change the source CID of AJ1 CID+12169
CID+12169 was copied from the GSUB vert destination CID+7956
with rotated by 270 (-90) degrees.
However, CID+7608, another GSUB vert source of CID+7956,
is a glyph with almost the same shape as the rotated glyph.
So this commit changes the source of CID+12169 to CID+7608
and stops rotating.
Both glyphs before and after this change
do not look out of place alongside CID+12170,
which is used in pairs with CID+12169.